refactor(ui): migrate admin, problem words, and settings pages to feature containers
- Extract admin management UI into AdminsContainer component under features/admins
- Extract problem words UI into ProblemWordsContainer component under features/problem-words
- Extract settings UI into SettingsContainer component under features/settings
- Remove inline component logic and UI, delegating rendering to respective container components
- Add metadata export for SEO titles and descriptions on pages
- Implement AdminTableSkeleton and ProblemWordsTableSkeleton for loading states
- Introduce reusable form components ProfileForm and SecurityForm under features/settings/components
- Simplify page components to return feature containers only, reducing code duplication and improving maintainability
