+++ apps/admin/src/app/dashboard/settings/page.tsx
@@ -1,187 +1,11 @@
-'use client';
+import type { Metadata } from 'next';
+import { SettingsContainer } from '../../../features/settings';
 
-import {
-  Box,
-  Button,
-  Card,
-  CardBody,
-  CardHeader,
-  FormControl,
-  FormErrorMessage,
-  FormLabel,
-  Heading,
-  Input,
-  useToast,
-  VStack,
-} from '@chakra-ui/react';
-import { axiosInstance, useAuthStore } from '@ielts/auth';
-import { useMutation } from '@tanstack/react-query';
-import { useEffect } from 'react';
-import { useForm } from 'react-hook-form';
+export const metadata: Metadata = {
+  title: 'Settings - Admin Panel',
+  description: 'Manage your account settings and preferences.',
+};
 
 export default function SettingsPage() {
-  const { user } = useAuthStore();
-  const toast = useToast();
-
-  const {
-    register: registerProfile,
-    handleSubmit: handleProfileSubmit,
-    reset: resetProfile,
-    formState: { isSubmitting: isProfileSubmitting },
-  } = useForm({
-    defaultValues: {
-      name: user?.name || '',
-    },
-  });
-
-  useEffect(() => {
-    if (user) {
-      resetProfile({ name: user.name });
-    }
-  }, [user, resetProfile]);
-
-  const {
-    register: registerPassword,
-    handleSubmit: handlePasswordSubmit,
-    reset: resetPassword,
-    formState: { errors: passwordErrors, isSubmitting: isPasswordSubmitting },
-  } = useForm();
-
-  const updateProfileMutation = useMutation({
-    mutationFn: async (data: { name: string }) => {
-      const response = await axiosInstance.patch('/admin/profile', data);
-      return response.data;
-    },
-    onSuccess: (_response, variables) => {
-      // Update local auth store with new user data
-      if (user) {
-        useAuthStore.getState().setUser({ ...user, name: variables.name });
-      }
-      toast({ title: 'Profile updated successfully', status: 'success' });
-    },
-    onError: (error: any) => {
-      toast({
-        title: 'Failed to update profile',
-        description: error.response?.data?.message,
-        status: 'error',
-      });
-    },
-  });
-
-  const changePasswordMutation = useMutation({
-    mutationFn: async (data: any) => {
-      await axiosInstance.post('/admin/change-password', data);
-    },
-    onSuccess: () => {
-      toast({ title: 'Password changed successfully', status: 'success' });
-      resetPassword();
-    },
-    onError: (error: any) => {
-      toast({
-        title: 'Failed to change password',
-        description: error.response?.data?.message,
-        status: 'error',
-      });
-    },
-  });
-
-  const onProfileSubmit = (data: any) => {
-    updateProfileMutation.mutate(data);
-  };
-
-  const onPasswordSubmit = (data: any) => {
-    changePasswordMutation.mutate({
-      currentPassword: data.currentPassword,
-      newPassword: data.newPassword,
-    });
-  };
-
-  return (
-    <Box maxW="container.md" mx="auto">
-      <VStack spacing={8} align="stretch">
-        <Heading size="lg">Account Settings</Heading>
-
-        <Card>
-          <CardHeader>
-            <Heading size="md">Profile Information</Heading>
-          </CardHeader>
-          <CardBody>
-            <form onSubmit={handleProfileSubmit(onProfileSubmit)}>
-              <VStack spacing={4} align="start">
-                <FormControl>
-                  <FormLabel>Email</FormLabel>
-                  <Input value={user?.email} isReadOnly bg="gray.50" />
-                </FormControl>
-                <FormControl>
-                  <FormLabel>Name</FormLabel>
-                  <Input {...registerProfile('name', { required: true })} />
-                </FormControl>
-                <Button
-                  type="submit"
-                  colorScheme="brand"
-                  isLoading={updateProfileMutation.isPending}
-                >
-                  Save Changes
-                </Button>
-              </VStack>
-            </form>
-          </CardBody>
-        </Card>
-
-        <Card>
-          <CardHeader>
-            <Heading size="md">Change Password</Heading>
-          </CardHeader>
-          <CardBody>
-            <form onSubmit={handlePasswordSubmit(onPasswordSubmit)}>
-              <VStack spacing={4} align="start">
-                <FormControl isRequired>
-                  <FormLabel>Current Password</FormLabel>
-                  <Input
-                    type="password"
-                    {...registerPassword('currentPassword', { required: true })}
-                  />
-                </FormControl>
-                <FormControl isRequired isInvalid={!!passwordErrors.newPassword}>
-                  <FormLabel>New Password</FormLabel>
-                  <Input
-                    type="password"
-                    {...registerPassword('newPassword', {
-                      required: true,
-                      minLength: { value: 6, message: 'Minimum 6 characters' },
-                    })}
-                  />
-                  <FormErrorMessage>
-                    {passwordErrors.newPassword?.message as string}
-                  </FormErrorMessage>
-                </FormControl>
-                <FormControl isRequired isInvalid={!!passwordErrors.confirmPassword}>
-                  <FormLabel>Confirm New Password</FormLabel>
-                  <Input
-                    type="password"
-                    {...registerPassword('confirmPassword', {
-                      required: true,
-                      validate: (value, formValues) =>
-                        value === formValues.newPassword || 'Passwords do not match',
-                    })}
-                  />
-                  <FormErrorMessage>
-                    {passwordErrors.confirmPassword &&
-                      (passwordErrors.confirmPassword.message as string)}
-                  </FormErrorMessage>
-                </FormControl>
-                <Button
-                  type="submit"
-                  colorScheme="brand"
-                  isLoading={changePasswordMutation.isPending}
-                >
-                  Update Password
-                </Button>
-              </VStack>
-            </form>
-          </CardBody>
-        </Card>
-      </VStack>
-    </Box>
-  );
+  return <SettingsContainer />;
 }
