+++ apps/admin/src/features/settings/types.ts
@@ -1,0 +1,10 @@
+export interface SettingsUser {
+  name: string;
+  email: string;
+}
+
+export interface PasswordFormData {
+  currentPassword: string;
+  newPassword: string;
+  confirmPassword?: string;
+}
